pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["kk71640 <kk71640@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21.0"
pandas = "^1.3.0"
matplotlib = "^3.4.2"
seaborn = "^0.11.1"
sklearn = "^0.0"
torch = "^1.9.0"
torchvision = "^0.10.0"
Pillow = "^8.3.0"
opencv-python = "^4.5.2"
albumentations = "^1.0.0"
tqdm = "^4.61.1"
pretrainedmodels = "^0.7.4"
notebook = "^6.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
